Move IndexList into Clockface directory
