+++ ui/src/clockface/components/index_views/IndexList.scss
