+++ ui/src/clockface/components/index_views/IndexListRow.tsx
