+++ ui/src/clockface/components/index_views/IndexListRowCell.tsx
